+++ src/models/types.ts
@@ -2,7 +2,8 @@ import { BooleanColumn, NumberColumn, TextColumn, DateColumn } from '../columns'
 import { AggregateFunction } from '../AggregateFunction'
 import { Expression } from './Expression'
 
-export type PrimitiveType = null|boolean|number|string|Date
+export type NonNullPrimitiveType = boolean|number|string|Date
+export type PrimitiveType = null|NonNullPrimitiveType
 
 const booleanArray: readonly string[] = ['t', 'tr', 'tru', 'true', 'f', 'fa', 'fal', 'fals', 'false']
 type TextBooleanSmallLetter = typeof booleanArray[number]
@@ -12,7 +13,8 @@ export type BooleanLike = boolean|TextBoolean|BooleanColumn
 export type NumberLike = number|NumberColumn
 export type TextLike = string|TextColumn
 export type DateLike = Date|DateColumn
-export type ValueType = null|BooleanLike|NumberLike|TextLike|DateLike
+export type ValueLike = BooleanLike|NumberLike|TextLike|DateLike
+export type ValueType = null|ValueLike
 export type OperandType = ValueType|AggregateFunction|Expression
 
 export function isTextBoolean(text: unknown): text is TextBoolean {